NCache Docker file and scripts changes for 5.3.2

+++ enterprise/java-edition/client/Linux/resources/installncache.sh
@@ -41,12 +41,11 @@ while [ "$1" != "" ]; do
 			VERBOSE="true"
 			;;
 
-	-h | --help )
-		
+	-h | --help )		
 			exit
 			;;
-	* )
-	
+
+	* )	
 			exit 1
     esac
     shift
@@ -62,22 +61,27 @@ then
 	PASSWORD="ncache"
 fi
 
+setup_file="ncache.ent.java.tar.gz"
+
+if [ ! -e "$setup_file"  ]; then
+    echo "Error: $setup_file does not exist."
+	exit 1
+fi
+
 #--- Untaring and installing NCache
-tar -zxf ncache.ent.net.tar.gz
+tar -zxf $setup_file
 cd ncache-enterprise
 
-./install --firstname $FIRST_NAME --lastname $LAST_NAME --email $EMAIL --company $COMPANY --installmode $INSTALLMODE --installkey $KEY
+./install --firstname $FIRST_NAME --lastname $LAST_NAME --email $EMAIL --company $COMPANY --installmode $INSTALLMODE
 
 cd ..
 #--- Removing installation resources
-rm -f ncache.ent.net.tar.gz
+rm -f $setup_file
 rm -rf ncache-enterprise/
-
-mkdir /home/ncache
 
 #--- Updating permissions and ownership
 chmod -R 775 /home/ncache /opt/ncache/bin/tools/web /opt/ncache/bin/service /opt/ncache/libexec 
-chown -R ncache:root /app /opt/ncache /home/ncache
+chown -R ncache:root /app
 
 #--- Adding the user to the root group
 usermod -a -G root ncache
